fix: auth response models fixed

+++ lib/feature/auth/data/model/send_otp_response.dart
@@ -1,6 +1,6 @@
 class SendOtpResponse {
   String otpToken;
-  int cooldown;
+  double cooldown;
 
   SendOtpResponse({
     required this.otpToken,
